Removal of unboundChannelThread
Replaced temporary with a new naming convention, but later decided its an unneccesary thread.
Establishing connection of a channel that is waiting for a declaration of exchange can be done in the main consumption thread.
This is because, even at best implementation, it still cannot be parallel (due to ConnectionBase's mutex locking).  What ended up happening, was resource starvation due to the two threads fighting over connection.  So i made it full concurrent and threw it in main thread.

This also allows the ability to subscribe while running